% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dose_Combination_Code.R
\name{Est_eqrest}
\alias{Est_eqrest}
\title{\code{Est_eqrest} function estimating the mean under the null constraints for the bootstrap method}
\usage{
Est_eqrest(Y, model, ind, indcomb, len, levelsDrugA, levelsDrugB)
}
\arguments{
\item{Y}{obsservation for response}

\item{model}{Annova model fitted to the data}

\item{ind}{all the drug dose combination groups (index including monotherapies and placebo control) in our study}

\item{indcomb}{all the drug dose combination groups (index excluding monotherapies and placebo control) in our study}

\item{len}{Variable contain the number of patients in each dose combination group}

\item{levelsDrugA}{no of dose groups for Drug A (r)}

\item{levelsDrugB}{no of dose groups for Drug B (s)}
}
\value{
a vector
}
\description{
\code{Est_eqrest} function estimating the mean under the null constraints for the bootstrap method
}
\references{
\url{http://stackoverflow.com/a/5577647}
}
