% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dose_Combination_Code.R
\name{gr}
\alias{gr}
\title{\code{gr} function evaluating the gradient of the results from eval_f function with respect to the mean estimates
gr(x)}
\usage{
gr(x)
}
\description{
\code{gr} function evaluating the gradient of the results from eval_f function with respect to the mean estimates
gr(x)
}
